add UserDetailsController

+++ server/src/main/java/com/example/server/service/userdetails/UserDetailsRepository.java
@@ -3,5 +3,5 @@ package com.example.server.service.userdetails;
 import com.example.server.model.UserDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDetailsRepository extends JpaRepository<UserDetails, Long> {
+public interface UserDetailsRepository extends JpaRepository<UserDetails, Integer> {
 }
